+++ CHANGELOG.md
@@ -17,6 +17,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - A method to ModelTestCase for running basic model tests when you aren't using config files.
 
+### Added
+
+- Added an option to `file_utils.cached_path` to automatically extract archives.
+- Added the ability to pass an archive file instead of a local directory to `Vocab.from_files`.
+- Added the ability to pass an archive file instead of a glob to `ShardedDatasetReader`.
+
 ## [v1.0.0](https://github.com/allenai/allennlp/releases/tag/v1.0.0) - 2020-06-16
 
 ### Fixed